nyota-enterprise/revenue/requirements.txt
fastapi
uvicorn
nats-py
pydantic
psycopg2-binary
httpx
openai
